Extract ProgrammingExercise styling and also apply to in-browser editor
