+++ src/partials/ProgrammingExercise/index.js
@@ -1,11 +1,7 @@
 import React from "react"
 import styled from "styled-components"
-import ContentLoader from "react-content-loader"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faPencilAlt as icon, faRedo } from "@fortawesome/free-solid-svg-icons"
 import { OutboundLink } from "gatsby-plugin-google-analytics"
 import { get } from "lodash"
-import { Card, CardContent, Button } from "@material-ui/core"
 
 import { withTranslation } from "react-i18next"
 import {
@@ -15,77 +11,10 @@ import {
 import LoginStateContext from "../../contexes/LoginStateContext"
 import LoginControls from "../../components/LoginControls"
 import withSimpleErrorBoundary from "../../util/withSimpleErrorBoundary"
-import { normalizeExerciseId } from "../../util/strings"
 import ExerciseDescription from "./ExerciseDescription"
 import ExtraDetails from "./ExtraDetails"
 import StyledDivider from "../../components/StyledDivider"
-
-const accentColor = "#FAAA38"
-
-const ProgrammingExerciseWrapper = styled(Card)`
-  margin: 3.5rem 0;
-  // border-left: 0.2rem solid ${accentColor};
-  border-radius: 1rem !important;
-  box-shadow: 0 8px 40px -12px rgba(0,0,0,0.3) !important;
-  padding: 0 !important;
-`
-
-const StyledIcon = styled(FontAwesomeIcon)`
-  vertical-align: middle;
-  margin-right: 1.5rem;
-  margin-left: 0.5rem;
-  color: white;
-  position: relative;
-  bottom: -13px;
-`
-
-const StyledRefreshIcon = styled(FontAwesomeIcon)`
-  color: white;
-`
-
-const Header = styled.div`
-  font-size: 1.3rem;
-  font-weight: normal;
-  padding 1rem 0;
-  border-bottom: 1px solid #f7f7f9;
-  background-color: #D23D48;
-  display: flex;
-  flex-direction: row;
-  align-items: 0;
-  color: white;
-  padding: 1rem;
-  padding-bottom: 1.5rem;
-  h3 {
-    margin-bottom: 0;
-  }
-`
-
-const HeaderTitleContainer = styled.div`
-  flex: 1;
-`
-
-const HeaderMuted = styled.span`
-  font-size: 18px;
-  font-weight: 400;
-  margin-right: 0.2rem;
-  position: relative;
-  bottom: -3px;
-`
-
-const PointsLabel = styled.span`
-  font-size: 18px;
-  font-weight: 400;
-`
-
-const PointContentWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`
-
-const Body = styled.div`
-  padding-bottom: 0.5rem;
-  min-height: 300px;
-`
+import ProgrammingExerciseCard from "./ProgrammingExerciseCard"
 
 const LoginNag = styled.div`
   margin-bottom: 1rem;
@@ -98,27 +27,11 @@ const LoginNagWrapper = styled.div`
   justify-content: center;
 `
 
-const PointsWrapper = styled.div`
-  margin-left: 0.5rem;
-  display: flex;
-  flex-direction: column;
-  text-align: right;
-  color: white;
-`
-
 const Small = styled.div`
   p {
     font-size: 0.9rem;
     color: #333;
   }
-`
-
-const StyledQuizPointsContentLoader = styled(ContentLoader)`
-  width: 100%;
-  max-width: 30px;
-  height: 31.2px;
-  position: relative;
-  top: -4px;
 `
 
 class ProgrammingExercise extends React.Component {
@@ -213,110 +126,47 @@ class ProgrammingExercise extends React.Component {
     )
 
     return (
-      <ProgrammingExerciseWrapper
-        id={normalizeExerciseId(`programming-exercise-${name}`)}
+      <ProgrammingExerciseCard
+        name={name}
+        points={points}
+        awardedPoints={awardedPoints}
+        onRefresh={this.onUpdate}
+        allowRefresh={this.context.loggedIn}
       >
-        <Header>
-          <StyledIcon icon={icon} size="2x" />
-          <HeaderTitleContainer>
-            <HeaderMuted>{this.props.t("programmingExercise")} </HeaderMuted>
-            <h3>{name}</h3>
-          </HeaderTitleContainer>
-
-          {this.context.loggedIn && (
-            <Button onClick={this.onUpdate}>
-              <StyledRefreshIcon icon={faRedo} />
-            </Button>
-          )}
-          <PointsWrapper>
-            <PointsLabel>{this.props.t("points")}</PointsLabel>
-
-            <PointContentWrapper>
-              {awardedPoints !== undefined ? (
-                <span>{awardedPoints}</span>
-              ) : (
-                <StyledQuizPointsContentLoader
-                  animate={!points}
-                  height={40}
-                  width={30}
-                  speed={2}
-                  primaryColor="#ffffff"
-                  primaryOpacity={0.6}
-                  secondaryColor="#dddddd"
-                  secondaryOpacity={0.6}
-                >
-                  <rect x="0" y="10" rx="12" ry="12" width="30" height="30" />
-                </StyledQuizPointsContentLoader>
-              )}
-              <span>/</span>
-              {points ? (
-                <span>{points}</span>
-              ) : (
-                <StyledQuizPointsContentLoader
-                  animate
-                  height={40}
-                  width={30}
-                  speed={2}
-                  primaryColor="#ffffff"
-                  primaryOpacity={0.6}
-                  secondaryColor="#dddddd"
-                  secondaryOpacity={0.6}
-                >
-                  <rect x="0" y="10" rx="12" ry="12" width="30" height="30" />
-                </StyledQuizPointsContentLoader>
-              )}
-            </PointContentWrapper>
-          </PointsWrapper>
-        </Header>
-        <CardContent>
-          <Body>
+        <div>
+          {this.context.loggedIn ? (
             <div>
-              {this.context.loggedIn ? (
-                <div>
-                  {points && points > 1 && (
-                    <Small>
-                      <p>
-                        {this.props.t("submitNB")}{" "}
-                        <OutboundLink
-                          href="https://www.mooc.fi/fi/installation/netbeans"
-                          rel="noopener noreferrer"
-                          target="_blank"
-                        >
-                          {this.props.t("submitHowTo")}
-                        </OutboundLink>
-                        .
-                      </p>
-                      <StyledDivider />
-                    </Small>
-                  )}
-                  <ExerciseDescription>{children}</ExerciseDescription>
-                  {this.state.exerciseDetails === null && (
-                    <div>Error loading exercise details</div>
-                  )}
-                </div>
-              ) : (
-                <div>
-                  <LoginNag>{this.props.t("loginForExercise")}</LoginNag>
-                  <LoginNagWrapper>
-                    <LoginControls />
-                  </LoginNagWrapper>
-                </div>
+              {points && points > 1 && (
+                <Small>
+                  <p>
+                    {this.props.t("submitNB")}{" "}
+                    <OutboundLink
+                      href="https://www.mooc.fi/fi/installation/netbeans"
+                      rel="noopener noreferrer"
+                      target="_blank"
+                    >
+                      {this.props.t("submitHowTo")}
+                    </OutboundLink>
+                    .
+                  </p>
+                  <StyledDivider />
+                </Small>
+              )}
+              <ExerciseDescription>{children}</ExerciseDescription>
+              {this.state.exerciseDetails === null && (
+                <div>Error loading exercise details</div>
               )}
             </div>
-
-            {this.context.loggedIn && (
-              <div>
-                <StyledDivider />
-                <ExtraDetails
-                  exerciseDetails={this.state.exerciseDetails}
-                  onUpdate={this.onUpdate}
-                  nocoins={this.props.nocoins}
-                />
-              </div>
-            )}
-          </Body>
-        </CardContent>
-      </ProgrammingExerciseWrapper>
+          ) : (
+            <div>
+              <LoginNag>{this.props.t("loginForExercise")}</LoginNag>
+              <LoginNagWrapper>
+                <LoginControls />
+              </LoginNagWrapper>
+            </div>
+          )}
+        </div>
+      </ProgrammingExerciseCard>
     )
   }
 }
